category usage validation

+++ src/main/java/cz/muni/fi/pv168/project/business/service/crud/CategoryCrudService.java
@@ -2,27 +2,34 @@ package cz.muni.fi.pv168.project.business.service.crud;
 
 import cz.muni.fi.pv168.project.business.model.Category;
 import cz.muni.fi.pv168.project.business.model.GuidProvider;
+import cz.muni.fi.pv168.project.business.model.UuidGuidProvider;
 import cz.muni.fi.pv168.project.business.repository.Repository;
-import cz.muni.fi.pv168.project.business.service.validation.DuplicateValidator;
-import cz.muni.fi.pv168.project.business.service.validation.ValidationResult;
-import cz.muni.fi.pv168.project.business.service.validation.Validator;
+import cz.muni.fi.pv168.project.business.service.validation.*;
+import cz.muni.fi.pv168.project.storage.DataStorageException;
+import cz.muni.fi.pv168.project.storage.sql.CategorySqlRepository;
+import cz.muni.fi.pv168.project.ui.MainWindow;
 
 import java.util.List;
 
 public class CategoryCrudService implements CrudService<Category> {
 
+
     private final Repository<Category> categoryRepository;
     private final Validator<Category> categoryValidator;
+    private final Validator<Category> usageValidator;
     private final GuidProvider guidProvider;
+
 
     public CategoryCrudService(
             Repository<Category> categoryRepository,
             Validator<Category> categoryValidator,
-            GuidProvider guidProvider
+            CategoryUsageValidator usageValidator
     ) {
         this.categoryRepository = categoryRepository;
-        this.categoryValidator = categoryValidator.and(new DuplicateValidator<>(categoryRepository));
-        this.guidProvider = guidProvider;
+        this.categoryValidator = categoryValidator
+                .and(new DuplicateValidator<>(this.categoryRepository));
+        this.guidProvider = new UuidGuidProvider();
+        this.usageValidator = usageValidator;
     }
 
     @Override
@@ -60,8 +67,16 @@ public class CategoryCrudService implements CrudService<Category> {
     }
 
     @Override
-    public void deleteByGuid(String guid) {
-        categoryRepository.deleteByGuid(guid);
+    public ValidationResult deleteByGuid(String guid) {
+        Category toDelete = ((CategorySqlRepository) categoryRepository).findByGuid(guid)
+                .orElseThrow(
+                        () -> new DataStorageException("Category with guid: " + guid + "not found!")
+                );
+        ValidationResult validationResult = usageValidator.validate(toDelete);
+        if (validationResult.isValid()) {
+            categoryRepository.deleteByGuid(guid);
+        }
+        return validationResult;
     }
 
     @Override
